+++ resources/views/mail/text.blade.php
@@ -1,1 +1,5 @@
 {{ $textContent }}
+@if (! empty($unsubscribeUrl))
+
+{{ __('marketing::public.unsubscribe_text') }}: {{ $unsubscribeUrl }}
+@endif